Move shared packet_trimming helpers to tests/common to fix cross-feature dep
tests/tam/test_tam_mod.py imported create_blocking_scheduler, ConfigTrimming,
get_interface_peer_addresses, get_queue_trim_counters_json, and DEFAULT_DSCP
from tests/packet_trimming/, which violates the cross-feature import rule
enforced by Pre_test Dependency Check.

Move the shared scheduler/queue-counter helpers, the ConfigTrimming context
manager, and the related constants to tests/common/helpers/packet_trimming.py.
Both packet_trimming and tam now import from this common location.
packet_trimming/constants.py and packet_trimming_helper.py re-export the
moved names so internal callers continue to resolve them via bare names.

No behavioral changes to the moved code; signatures preserved verbatim.
